creating new reducer for singleUserOrders
